Added a new resource
Hey, I have added a new resource for loops, and for this, I gave a link for Java loops. I found this tutorial online while looking for some resources to learn loops in java. Topics were explained pretty well and found it to be useful and thought of adding it here. Hope this small contribution will help other learners.